Convert a character vector to a numeric integer for setting a reproducible seed

char2seed() is a parlor trick for converting a character vector into an integer for the sake of setting a reproducible seed.

char2seed(x)

Arguments

x
a character vector

Value

char2seed() takes a character vector and returns a reproducible seed for you to use for whatever purpose. It's used internally in codename().

Details

Interested users can see how this works. Namely, letters (and numbers) in the character vector are assigned corresponding numbers. These numbers are added together and then exponentiated to create a really big number. The number is divided over 1 minus the absolute maximum number that R can handle by default (2^32). The remainder of this division is what ultimately becomes the reproducible seed. The nature of this parlor trick means there's a possibility, however infinitesimally small, that two different character vectors can return the same reproducible seed. The function may warn of "loss of accuracy", but this just means you supplied it a really long character vector.

Author

Steven V. Miller

Examples

char2seed("ABCDEF")

#> [1] 636

char2seed("Go Bucks!")

#> [1] 1733

char2seed("My Project Title")

#> [1] 1902490